update the dist folder
