--- src/modules/index.js
@@ -1,17 +1,0 @@
-// Select element step 1
-const form = document.getElementById('todoform');
-const todoInput = document.getElementById('newtodo');
-// Let variables step 4 cretae array first
-const todos = [];
-
-// define SaveTo function do step 3
-function saveTodo() {
-
-}
-
-// Form submit step 2
-form.addEventListener('submit', (event) => {
-  event.preventDefault();
-
-  saveTodo();
-});
